recycle PVs at the end of the each test
