+++ tests/e2e/e2e.go
@@ -30,7 +30,6 @@ import (
 	"github.com/onsi/gomega"
 	asclientset "github.com/pingcap/advanced-statefulset/client/client/clientset/versioned"
 	"github.com/pingcap/tidb-operator/pkg/client/clientset/versioned"
-	"github.com/pingcap/tidb-operator/pkg/label"
 	"github.com/pingcap/tidb-operator/pkg/version"
 	"github.com/pingcap/tidb-operator/tests"
 	e2econfig "github.com/pingcap/tidb-operator/tests/e2e/config"
@@ -40,7 +39,6 @@ import (
 	v1 "k8s.io/api/core/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	runtimeutils "k8s.io/apimachinery/pkg/util/runtime"
@@ -271,58 +269,6 @@ var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
 		return true, nil
 	})
 	framework.ExpectNoError(err, "failed to wait for all PVs to be available")
-
-	// tidb-operator will set persistentVolumeReclaimPolicy to Retain if users
-	// reqeust this. To reduce storage usage, we set
-	// persistentVolumeReclaimPolicy to Delete if the PVC namespace is gone.
-	go wait.Forever(func() {
-		pvList, err := kubeCli.CoreV1().PersistentVolumes().List(metav1.ListOptions{})
-		if err != nil {
-			framework.Logf("failed to list pvs: %v", err)
-			return
-		}
-		var (
-			total          int = len(pvList.Items)
-			retainReleased int
-			skipped        int
-			failed         int
-			succeeded      int
-		)
-		defer func() {
-			framework.Logf("recycling orphan PVs (total: %d, retainReleased: %d, skipped: %d, failed: %d, succeeded: %d)", total, retainReleased, skipped, failed, succeeded)
-		}()
-		for _, pv := range pvList.Items {
-			if pv.Spec.PersistentVolumeReclaimPolicy != v1.PersistentVolumeReclaimRetain || pv.Status.Phase != v1.VolumeReleased {
-				continue
-			}
-			retainReleased++
-			pvcNamespaceName, ok := pv.Labels[label.NamespaceLabelKey]
-			if !ok {
-				framework.Logf("label %q does not exist in PV %q", label.NamespaceLabelKey, pv.Name)
-				failed++
-				continue
-			}
-			_, err := kubeCli.CoreV1().Namespaces().Get(pvcNamespaceName, metav1.GetOptions{})
-			if err != nil && !apierrors.IsNotFound(err) {
-				framework.Logf("failed to get namespace %q: %v", pvcNamespaceName, err)
-				failed++
-				continue
-			}
-			if apierrors.IsNotFound(err) {
-				skipped++
-				continue
-			}
-			pv.Spec.PersistentVolumeReclaimPolicy = v1.PersistentVolumeReclaimDelete
-			_, err = kubeCli.CoreV1().PersistentVolumes().Update(&pv)
-			if err != nil {
-				failed++
-				framework.Logf("failed to set PersistentVolumeReclaimPolicy of PV %q to Delete: %v", pv.Name, err)
-			} else {
-				succeeded++
-				framework.Logf("successfully set PersistentVolumeReclaimPolicy of PV %q to Delete", pv.Name)
-			}
-		}
-	}, time.Second*10)
 
 	ginkgo.By("Labeling nodes")
 	oa := tests.NewOperatorActions(cli, kubeCli, asCli, aggrCli, apiExtCli, tests.DefaultPollInterval, nil, e2econfig.TestConfig, nil, nil, nil)
